requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pandas==2.2.3
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
openpyxl==3.1.5
psycopg2-binary==2.9.9
python-dotenv==1.0.1
streamlit==1.37.1